+++ app/Http/Controllers/PelangganController.php
@@ -50,14 +50,49 @@ class PelangganController extends Controller
     //detail pelanggan
     public function show($id, Request $request)
     {
-        $pelanggan = Pelanggan::where('id_users', $request->user()->id)->findOrFail($id);
+        // $pelanggan = Pelanggan::where('id_users', $request->user()->id)->findOrFail($id);
+        // return response()->json($pelanggan, 200);
+
+        $pelanggan = Pelanggan::find($id);
+
+        // jika id tidak ada
+        if (!$pelanggan) {
+            return response()->json([
+                'message' => 'Pelanggan ID tidak ditemukan'
+            ], 404);
+        }
+
+        // cek token
+        if ($pelanggan->id_users !== $request->user()->id) {
+            return response()->json([
+                'message' => 'Token invalid'
+            ], 400);
+        }
+
         return response()->json($pelanggan, 200);
     }
 
     //update pelanggan
     public function update(Request $request, $id)
     {
-        $pelanggan = Pelanggan::where('id_users', $request->user()->id)->findOrFail($id);
+        // $pelanggan = Pelanggan::where('id_users', $request->user()->id)->findOrFail($id);
+
+        $pelanggan = Pelanggan::find($id);
+
+        // tidak ada id
+        if (!$pelanggan) {
+            return response()->json([
+                'message' => 'Pelanggan ID tidak ditemukan'
+            ], 404);
+        }
+
+        // cek token
+        if ($pelanggan->id_users !== $request->user()->id) {
+            return response()->json([
+                'message' => 'Token invalid'
+            ], 400);
+        }
+
 
         $request->validate([
             'email_pelanggan' => 'sometimes|email|unique:pelanggans,email_pelanggan,' . $pelanggan->id,
@@ -79,7 +114,24 @@ class PelangganController extends Controller
     //hapus pelanggan
     public function destroy($id, Request $request)
     {
-        $pelanggan = Pelanggan::where('id_users', $request->user()->id)->findOrFail($id);
+        // $pelanggan = Pelanggan::where('id_users', $request->user()->id)->findOrFail($id);
+
+        $pelanggan = Pelanggan::find($id);
+        
+        // jika id tidak ada
+        if (!$pelanggan) {
+            return response()->json([
+                'message' => 'Pelanggan ID tidak ditemukan'
+            ], 404);
+        }
+
+        // cek token
+        if ($pelanggan->id_users !== $request->user()->id) { //harusnya cek ke api_token
+            return response()->json([
+                'message' => 'Token invalid'
+            ], 400);
+        }
+
         $pelanggan->delete();
 
         return response()->json([
